move rules & checks tests
