+++ src/tests/unit/tests/scanner/custom-rules/css-positioning-rule.test.ts
@@ -3,8 +3,8 @@
 import * as axe from 'axe-core';
 import { GlobalMock, GlobalScope, It, MockBehavior, Times } from 'typemoq';
 
-import { cssPositioningConfiguration } from '../../../../scanner/custom-rules/css-positioning-rule';
-import { DictionaryStringTo } from '../../../../types/common-types';
+import { cssPositioningConfiguration } from '../../../../../scanner/custom-rules/css-positioning-rule';
+import { DictionaryStringTo } from '../../../../../types/common-types';
 
 describe('verify meaningful sequence configs', () => {
     it('should have correct props', () => {
