+++ src/tests/unit/tests/scanner/custom-rules/custom-widget-check.test.ts
@@ -4,8 +4,8 @@ import * as Axe from 'axe-core';
 import { difference, map } from 'lodash';
 import { Mock, MockBehavior } from 'typemoq';
 
-import { customWidgetConfiguration } from '../../../../scanner/custom-rules/custom-widget';
-import { ICheckConfiguration } from '../../../../scanner/iruleresults';
+import { customWidgetConfiguration } from '../../../../../scanner/custom-rules/custom-widget';
+import { ICheckConfiguration } from '../../../../../scanner/iruleresults';
 
 const axe = Axe as any;
 const fixture = createTestFixture('test-fixture', '');
