+++ src/tests/unit/tests/scanner/custom-rules/landmark-role.test.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { landmarkConfiguration } from '../../../../scanner/custom-rules/landmark-rule';
+import { landmarkConfiguration } from '../../../../../scanner/custom-rules/landmark-rule';
 
 describe('landmarkRule', () => {
     describe('verify landmark configs', () => {
