+++ src/tests/unit/tests/scanner/custom-rules/link-function.test.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 import { GlobalMock, GlobalScope, IGlobalMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
-import * as AxeUtils from '../../../../scanner/axe-utils';
-import { linkFunctionConfiguration } from '../../../../scanner/custom-rules/link-function';
+import * as AxeUtils from '../../../../../scanner/axe-utils';
+import { linkFunctionConfiguration } from '../../../../../scanner/custom-rules/link-function';
 
 const outerHTML = 'outerHTML';
 const parentOuterHTML = 'parentOuterHTML';
