+++ src/tests/unit/tests/scanner/custom-rules/text-alternative.test.ts
@@ -3,8 +3,8 @@
 import * as Axe from 'axe-core';
 import { GlobalMock, GlobalScope, IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
-import * as AxeUtils from '../../../../scanner/axe-utils';
-import { textAlternativeConfiguration } from '../../../../scanner/custom-rules/text-alternative';
+import * as AxeUtils from '../../../../../scanner/axe-utils';
+import { textAlternativeConfiguration } from '../../../../../scanner/custom-rules/text-alternative';
 
 describe('text alternative', () => {
     describe('verify text alternative configs', () => {
