+++ src/tests/unit/tests/scanner/custom-rules/text-contrast.test.ts
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 import { GlobalMock, GlobalScope, IGlobalMock, IMock, It, Mock, MockBehavior } from 'typemoq';
 
-import * as AxeUtils from '../../../../scanner/axe-utils';
-import { textContrastConfiguration } from '../../../../scanner/custom-rules/text-contrast';
-import { DictionaryStringTo } from '../../../../types/common-types';
+import * as AxeUtils from '../../../../../scanner/axe-utils';
+import { textContrastConfiguration } from '../../../../../scanner/custom-rules/text-contrast';
+import { DictionaryStringTo } from '../../../../../types/common-types';
 
 function testTextContrast(
     node: DictionaryStringTo<string>,
