+++ src/tests/unit/tests/scanner/custom-rules/widget-function.test.ts
@@ -3,9 +3,9 @@
 import { difference, map } from 'lodash';
 import { GlobalMock, GlobalScope, It, Mock, MockBehavior, Times } from 'typemoq';
 
-import * as AxeUtils from '../../../../scanner/axe-utils';
-import { evaluateWidgetFunction, widgetFunctionConfiguration } from '../../../../scanner/custom-rules/widget-function';
-import { createNodeStub, testNativeWidgetConfiguration } from './helpers';
+import * as AxeUtils from '../../../../../scanner/axe-utils';
+import { evaluateWidgetFunction, widgetFunctionConfiguration } from '../../../../../scanner/custom-rules/widget-function';
+import { createNodeStub, testNativeWidgetConfiguration } from '../helpers';
 
 declare let axe;
